requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
joblib
reportlab
python-pptx
openpyxl
